% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.well.R
\name{is.well}
\alias{is.well}
\title{Determine Whether a Given String is a Valid Well Label}
\usage{
is.well(well)
}
\arguments{
\item{well}{A well label (e.g., \code{"A6"})}
}
\value{
A logical value indicating whether or not the given well label is
valid
}
\description{
Determine Whether a Given String is a Valid Well Label
}
\examples{
is.well("A6")
is.well(c("A3", "A4", "G6"))

}
