% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtiter_data_frame.R
\name{microtiter_data_frame}
\alias{microtiter_data_frame}
\alias{microtiter_tibble}
\title{Create a Data Frame or Tibble with All Wells in a Microtiter Plate}
\usage{
microtiter_data_frame(nwells = 96, include_well = TRUE)

microtiter_tibble(nwells = 96, include_well = TRUE)
}
\arguments{
\item{nwells}{Microtiter plate format specified as number of wells (e.g., 6 or 24; default: 96)}

\item{include_well}{Logical value indicating whether or not to include
well labels (default: \code{TRUE})}
}
\value{
A data frame
}
\description{
\code{microtiter_data_frame} creates a data frame representing a microtiter
plate, with one row per well, and columns for Row, Column, and Well
(optional). \code{microtiter_tibble} creates a \link[tibble]{tibble} instead.
}
\examples{
microtiter_data_frame(nwells = 96)

}
