% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_letter.R
\name{row_letter}
\alias{row_letter}
\title{Get the letter associated with the given row number(s)}
\usage{
row_letter(rownum)
}
\arguments{
\item{rownum}{The row number of a well}
}
\value{
A character containing the row letter
}
\description{
TODO
}
\note{
This currently does not fully support 1536-well plates
}
\examples{
row_letter(3)

}
