% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_letters.R
\name{row_letters}
\alias{row_letters}
\title{Get the letters associated with the given row number(s)}
\usage{
row_letters(rownum)
}
\arguments{
\item{rownum}{The row number of a well}
}
\value{
A character containing the row letter
}
\description{
TODO
}
\note{
This currently does not fully support 1536-well plates
}
\examples{
row_letters(3)

}
