% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_column.R
\name{well_column}
\alias{well_column}
\title{Get column number from a well label}
\usage{
well_column(well)
}
\arguments{
\item{well}{A well label (e.g., "A4")}
}
\value{
Integer value for the corresponding column
}
\description{
TODO
}
\examples{
well_column("C6")

}
\seealso{
\code{\link{well_row}}
}
