% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_label.R
\name{well_label}
\alias{well_label}
\title{Get a Well Label for the Given Row and Column}
\usage{
well_label(row, column)
}
\arguments{
\item{row}{Numeric row index(es)}

\item{column}{Numeric column index(es)}
}
\value{
String containing the corresponding Well Label
}
\description{
\code{well_label} returns a well label (e.g., "A4") for a given row and
column on a microtiter plate.
}
\examples{
well_label(1, 5)
well_label(1:8, 4)

}
