% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/well_row.R
\name{well_row}
\alias{well_row}
\title{Get row number from a well label}
\usage{
well_row(well)
}
\arguments{
\item{well}{A well label (e.g., "A4")}
}
\value{
Integer value for the corresponding row
}
\description{
TODO
}
\examples{
well_row("C6")

}
\seealso{
\code{\link{well_column}}
}
